requirements.txt
mysql-connector-python==8.0.31
python-dotenv==0.21.0
matplotlib==3.6.2